[GPUP] Use standard int type in RemoteMediaPlayerState https://bugs.webkit.org/show_bug.cgi?id=208962
Reviewed by Eric Carlson.

An unsigned long long is at least 64-bits so use uint64_t instead. This issue
was found when compiling out the GPU Process. No specialization of unsigned long
long was found which resulted in errors. This appears to be a non-Cocoa problem.

* WebProcess/GPU/media/RemoteMediaPlayerState.h:
(WebKit::RemoteMediaPlayerState::decode):


Canonical link: https://commits.webkit.org/221893@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258323 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebKit/WebProcess/GPU/media/RemoteMediaPlayerState.h
@@ -53,7 +53,7 @@ struct RemoteMediaPlayerState {
     double minFastReverseRate { 0 };
     double seekableTimeRangesLastModifiedTime { 0 };
     double liveUpdateInterval { 0 };
-    unsigned long long totalBytes { 0 };
+    uint64_t totalBytes { 0 };
     Optional<bool> wouldTaintDocumentSecurityOrigin { true };
     bool paused { true };
     bool loadingProgressed { false };
@@ -192,7 +192,7 @@ struct RemoteMediaPlayerState {
         if (!liveUpdateInterval)
             return WTF::nullopt;
 
-        Optional<unsigned long long> totalBytes;
+        Optional<uint64_t> totalBytes;
         decoder >> totalBytes;
         if (!totalBytes)
             return WTF::nullopt;
